login/reg with sanctum auth

+++ app/Http/Controllers/PostController.php
@@ -3,18 +3,27 @@
 namespace App\Http\Controllers;
 
 use App\Models\Post;
-// use App\Http\Requests\StorePostRequest;
-// use App\Http\Requests\UpdatePostRequest;
 use Illuminate\Http\Request;
+use Illuminate\Routing\Controllers\HasMiddleware;
+use Illuminate\Routing\Controllers\Middleware;
+use Illuminate\Support\Facades\Gate;
 
-class PostController extends Controller
+class PostController extends Controller implements HasMiddleware
 {
+    public static function middleware()
+    {
+        return [
+            new Middleware('auth:sanctum', except: ['index', 'show'])
+        ];
+    }
+
     /**
      * Display a listing of the resource.
      */
     public function index()
     {
-        return Post::all();
+        return Post::with('user')->latest()->get();
+        // return Post::all();
     }
 
     /**
@@ -27,9 +36,10 @@ class PostController extends Controller
             'body' => 'required'
         ]);
 
-        $post = Post::create($fields);
+        $post = $request->user()->posts()->create($fields);
 
-        return $post;
+        return ['post' => $post, 'user' => $post->user];
+        // return $post;
     }
 
     /**
@@ -37,7 +47,8 @@ class PostController extends Controller
      */
     public function show(Post $post)
     {
-        return $post;
+        return ['post' => $post, 'user' => $post->user];
+        // return $post;
     }
 
     /**
@@ -45,6 +56,8 @@ class PostController extends Controller
      */
     public function update(Request $request, Post $post)
     {
+        Gate::authorize('modify', $post);
+
         $fields = $request->validate([
             'title' => 'required|max:255',
             'body' => 'required'
@@ -52,7 +65,8 @@ class PostController extends Controller
 
         $post->update($fields);
 
-        return $post;
+        return ['post' => $post, 'user' => $post->user];
+        // return $post;
     }
 
     /**
@@ -60,8 +74,10 @@ class PostController extends Controller
      */
     public function destroy(Post $post)
     {
+        Gate::authorize('modify', $post);
+
         $post->delete();
 
-        return [ 'message' => 'The post was deleted'];
+        return ['message' => 'The post was deleted'];
     }
 }
